Some minor fixes based on MSVC

+++ src/PDFs/PdfBase.cu
@@ -157,7 +157,7 @@ __host__ void PdfBase::setData(DataSet *data) {
         numEntries = data->getNumEvents();
         numEvents  = numEntries;
 
-        int dimensions = observables.size();
+        size_t dimensions = observables.size();
 
 #ifdef GOOFIT_MPI
         // This fetches our rank and the total number of processes in the MPI call
